requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-extensions==2.2.5
django-model-utils==3.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.2.0
drf-yasg==1.17.1
entrypoints==0.3
environs==7.3.0
flake8==3.7.9
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
mccabe==0.6.1
model-mommy==2.0.0
packaging==20.3
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
